+++ src/engine/graphic/GRRitardando.cpp
@@ -80,7 +80,7 @@ GRRitardando::GRRitardando( GRStaff * stf, const ARRitardando * artrem )
   }
 
 //  mFontSize = artrem->getFSize(curLSPACE);
-  mFontSize = artrem->getFSize() * curLSPACE / LSPACE;
+  mFontSize = int(artrem->getFSize() * curLSPACE / LSPACE);
   if (mFontSize == 0)
 	  mFontSize = (int)(1.5f * LSPACE);
